arm64: smmu: pointer cast fix for tegra-smmu.c
Reviewed-on: http://git-master/r/276092
Tested-by: Hiroshi Doyu <hdoyu@nvidia.com>

+++ drivers/iommu/iommu.c
@@ -1024,7 +1024,7 @@ static size_t iommu_pgsize(struct iommu_domain *domain,
 }
 
 int iommu_map(struct iommu_domain *domain, unsigned long iova,
-	      phys_addr_t paddr, size_t size, int prot)
+	      phys_addr_t paddr, size_t size, unsigned long prot)
 {
 	unsigned long orig_iova = iova;
 	unsigned int min_pagesz;
@@ -1077,7 +1077,7 @@ int iommu_map(struct iommu_domain *domain, unsigned long iova,
 EXPORT_SYMBOL_GPL(iommu_map);
 
 int iommu_map_pages(struct iommu_domain *domain, unsigned long iova,
-		    struct page **pages, size_t count, int prot)
+		    struct page **pages, size_t count, unsigned long prot)
 {
 	int err;
 
@@ -1090,7 +1090,7 @@ int iommu_map_pages(struct iommu_domain *domain, unsigned long iova,
 EXPORT_SYMBOL_GPL(iommu_map_pages);
 
 int iommu_map_sg(struct iommu_domain *domain, unsigned long iova,
-		 struct scatterlist *sgl, int nents, int prot)
+		 struct scatterlist *sgl, int nents, unsigned long prot)
 {
 	int err;
